# Consolidate memory R/W operations