--- a/src/io.c
+++ b/src/io.c
@@ -5,65 +5,59 @@
 
 #include <assert.h>
 #include <stdint.h>
+#include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#if !defined(_WIN32)
+#define USE_MMAP 1
+#include <sys/mman.h>
+#include <unistd.h>
+#endif
 
 #include "io.h"
-#include "mpool.h"
 
-static const uint32_t mask_lo = 0xffff;
-static const uint32_t mask_hi = ~(0xffff);
-
-static struct mpool *mp;
+static uint8_t *data_memory_base;
+/* set memory size to 2^32 bytes */
+#define MEM_SIZE 0x100000000ULL
 
 memory_t *memory_new()
 {
-    memory_t *m = calloc(1, sizeof(memory_t));
-    /* Initialize the mpool size to sizeof(chunk_t) << 2, and it will extend
-     * automatically if needed */
-    mp = mpool_create(sizeof(chunk_t) << 2, sizeof(chunk_t));
-    return m;
+    memory_t *mem = malloc(sizeof(memory_t));
+#if defined(USE_MMAP)
+    data_memory_base = mmap(NULL, MEM_SIZE, PROT_READ | PROT_WRITE,
+                            MAP_ANONYMOUS | MAP_PRIVATE, -1, 0);
+    if (data_memory_base == MAP_FAILED)
+        return NULL;
+#else
+    data_memory_base = malloc(MEM_SIZE);
+    if (data_memory_base)
+        return NULL;
+#endif
+    mem->mem_base = data_memory_base;
+    mem->mem_size = MEM_SIZE;
+    return mem;
 }
 
-void memory_delete(memory_t *m)
+void memory_delete(memory_t *mem)
 {
-    if (!m)
-        return;
-    mpool_destory(mp);
-    free(m);
+    munmap(mem->mem_base, MEM_SIZE);
 }
 
-void memory_read(memory_t *m, uint8_t *dst, uint32_t addr, uint32_t size)
+void memory_read(memory_t *mem, uint8_t *dst, uint32_t addr, uint32_t size)
 {
-    /* test if this read is entirely within one chunk */
-    if ((addr & mask_hi) == ((addr + size) & mask_hi)) {
-        chunk_t *c;
-        if ((c = m->chunks[addr >> 16])) {
-            /* get the subchunk pointer */
-            const uint32_t p = (addr & mask_lo);
-
-            /* copy over the data */
-            memcpy(dst, c->data + p, size);
-        } else {
-            memset(dst, 0, size);
-        }
-    } else {
-        /* naive copy */
-        for (uint32_t i = 0; i < size; ++i) {
-            uint32_t p = addr + i;
-            chunk_t *c = m->chunks[p >> 16];
-            dst[i] = c ? c->data[p & 0xffff] : 0;
-        }
-    }
+    memcpy(dst, mem->mem_base + addr, size);
 }
 
-uint32_t memory_read_str(memory_t *m, uint8_t *dst, uint32_t addr, uint32_t max)
+uint32_t memory_read_str(memory_t *mem,
+                         uint8_t *dst,
+                         uint32_t addr,
+                         uint32_t max)
 {
     uint32_t len = 0;
     const uint8_t *end = dst + max;
     for (;; ++len, ++dst) {
         uint8_t ch = 0;
-        memory_read(m, &ch, addr + len, 1);
+        memory_read(mem, &ch, addr + len, 1);
         if (dst < end)
             *dst = ch;
         if (!ch)
@@ -72,76 +66,40 @@ uint32_t memory_read_str(memory_t *m, uint8_t *dst, uint32_t addr, uint32_t max)
     return len + 1;
 }
 
-uint32_t memory_ifetch(memory_t *m, uint32_t addr)
+uint32_t memory_ifetch(uint32_t addr)
 {
-    const uint32_t addr_lo = addr & mask_lo;
-    assert((addr_lo & 1) == 0);
-
-    chunk_t *c = m->chunks[addr >> 16];
-    assert(c);
-    return *(const uint32_t *) (c->data + addr_lo);
+    return *(const uint32_t *) (data_memory_base + addr);
 }
 
-uint32_t memory_read_w(memory_t *m, uint32_t addr)
-{
-    const uint32_t addr_lo = addr & mask_lo;
-    chunk_t *c = m->chunks[addr >> 16];
-    return *(const uint32_t *) (c->data + addr_lo);
-}
-
-uint16_t memory_read_s(memory_t *m, uint32_t addr)
-{
-    const uint32_t addr_lo = addr & mask_lo;
-    chunk_t *c = m->chunks[addr >> 16];
-    return *(const uint16_t *) (c->data + addr_lo);
-}
-
-uint8_t memory_read_b(memory_t *m, uint32_t addr)
-{
-    chunk_t *c = m->chunks[addr >> 16];
-    return c->data[addr & mask_lo];
-}
-
-void memory_write(memory_t *m, uint32_t addr, const uint8_t *src, uint32_t size)
-{
-    for (uint32_t i = 0; i < size; ++i) {
-        const uint32_t addr_lo = (addr + i) & mask_lo,
-                       addr_hi = (addr + i) >> 16;
-        chunk_t *c = m->chunks[addr_hi];
-        if (!c) {
-            c = mpool_calloc(mp);
-            m->chunks[addr_hi] = c;
-        }
-        c->data[addr_lo] = src[i];
+#define MEM_READ_IMPL(size, type)                   \
+    type memory_read_##size(uint32_t addr)          \
+    {                                               \
+        return *(type *) (data_memory_base + addr); \
     }
+
+MEM_READ_IMPL(w, uint32_t);
+MEM_READ_IMPL(s, uint16_t);
+MEM_READ_IMPL(b, uint8_t);
+
+void memory_write(memory_t *mem,
+                  uint32_t addr,
+                  const uint8_t *src,
+                  uint32_t size)
+{
+    memcpy(mem->mem_base + addr, src, size);
 }
 
-#define MEM_WRITE_IMPL(size, type)                                           \
-    void memory_write_##size(memory_t *m, uint32_t addr, const uint8_t *src) \
-    {                                                                        \
-        const uint32_t addr_lo = addr & mask_lo, addr_hi = addr >> 16;       \
-        chunk_t *c = m->chunks[addr_hi];                                     \
-        if (unlikely(!c)) {                                                  \
-            c = mpool_calloc(mp);                                            \
-            m->chunks[addr_hi] = c;                                          \
-        }                                                                    \
-        *(type *) (c->data + addr_lo) = *(const type *) src;                 \
+#define MEM_WRITE_IMPL(size, type)                                 \
+    void memory_write_##size(uint32_t addr, const uint8_t *src)    \
+    {                                                              \
+        *(type *) (data_memory_base + addr) = *(const type *) src; \
     }
 
 MEM_WRITE_IMPL(w, uint32_t);
 MEM_WRITE_IMPL(s, uint16_t);
 MEM_WRITE_IMPL(b, uint8_t);
 
-void memory_fill(memory_t *m, uint32_t addr, uint32_t size, uint8_t val)
+void memory_fill(memory_t *mem, uint32_t addr, uint32_t size, uint8_t val)
 {
-    for (uint32_t i = 0; i < size; ++i) {
-        uint32_t p = addr + i;
-        uint32_t x = p >> 16;
-        chunk_t *c = m->chunks[x];
-        if (!c) {
-            c = mpool_calloc(mp);
-            m->chunks[x] = c;
-        }
-        c->data[p & 0xffff] = val;
-    }
+    memset(mem->mem_base + addr, val, size);
 }