@qwe661234 2023-07-07
> Utilize [uftrace](https://github.com/namhyung/uftrace) tool to analyze function tracing.
> 
> Apply the following changes:
> 
> ```diff
> --- a/Makefile
> +++ b/Makefile
> @@ -7,6 +7,7 @@ BIN := $(OUT)/rv32emu
>  CFLAGS = -std=gnu99 -O2 -Wall -Wextra
>  CFLAGS += -Wno-unused-label
>  CFLAGS += -include src/common.h
> +CFLAGS += -g -pg
>  
>  # Set the default stack pointer
>  CFLAGS += -D DEFAULT_STACK_ADDR=0xFFFFF000
> @@ -80,7 +81,7 @@ endif
>  
>  # For tail-call elimination, we need a specific set of build flags applied.
>  # FIXME: On macOS + Apple Silicon, -fno-stack-protector might have a negative impact.
> -$(OUT)/emulate.o: CFLAGS += -fomit-frame-pointer -fno-stack-check -fno-stack-protector
> +$(OUT)/emulate.o: CFLAGS += -fno-stack-check -fno-stack-protector
>  
>  # Clear the .DEFAULT_GOAL special variable, so that the following turns
>  # to the first target after .DEFAULT_GOAL is not set.
> ```
> 
> And then `make clean all`.
> 
> Run
> 
> ```shell
> $ uftrace build/rv32emu build/hello.elf 
> ```
> 
> The reference output:
> 
> ```
>             [233860] |   memory_new() {
>   13.068 us [233860] |     calloc();
>             [233860] |     mpool_create() {
>    0.503 us [233860] |       malloc();
>    0.212 us [233860] |       getpagesize();
>    3.541 us [233860] |       mmap();
>   17.226 us [233860] |     } /* mpool_create */
>   31.291 us [233860] |   } /* memory_new */
> ```
> 
> You shall show the corresponding difference based on function tracing.

Based on the function tracing results from the ftrace tool, in the old memory design, `do_lw` takes approximately 9 us to allocate data memory and read data. Even when it doesn't require allocating data memory, it still takes around 1.7 us to read data. However, with the new design, we only need about 1 us to read data without allocating any data memory. Hence, the same applies to `do_sw`.
## sw
### old memory design
```
            [3659971] |    do_sw() {
            [3659971] |        on_mem_write_w() {
   0.255 us [3659971] |            rv_userdata();
   0.220 us [3659971] |            memory_write_w();
   9.141 us [3659971] |        } /* on_mem_write_w */
   ...
            [3659971] |    do_sw() {
            [3659971] |        on_mem_write_w() {
   0.265 us [3659971] |            rv_userdata();
   0.225 us [3659971] |            memory_write_w();
   1.795 us [3659971] |        } /* on_mem_write_w */
```
### new memory design
```
            [3658503] |    do_sw() {
            [3658503] |        on_mem_write_w() {
   0.245 us [3658503] |            memory_write_w();
   1.055 us [3658503] |        } /* on_mem_write_w */
```
### lw
### old memory design
```
            [3659971] |    do_lw() {
            [3659971] |        on_mem_read_w() {
   0.265 us [3659971] |            rv_userdata();
   0.255 us [3659971] |            memory_read_w();
   9.626 us [3659971] |         } /* on_mem_read_w */
   ...
            [3659971] |    do_lw() {
            [3659971] |        on_mem_read_w() {
   0.260 us [3659971] |            rv_userdata();
   0.235 us [3659971] |            memory_read_w();
   1.485 us [3659971] |        } /* on_mem_read_w */
```
### new memory design
```
       
            [3658503] |    do_lw() {
            [3658503] |        on_mem_read_w() {
   0.265 us [3658503] |            memory_read_w();
   1.025 us [3658503] |        } /* on_mem_read_w */
```