@jserv 2023-07-07
You shall refer the operating system by Apple to [macOS](https://en.wikipedia.org/wiki/MacOS) rather than MacOS. It is a matter of brand identity.