@jserv 2023-07-07
Always use the pair "original" vs. "proposed" while you are about to propose non-trivial changes by comprehensive comparisons.